Struct seedling_runtime::Permill

pub struct Permill(_);

A fixed point representation of a number in the range [0, 1].

Parts per Million

Implementations

impl Permill

pub const fn from_parts(parts: u32) -> Permill

From an explicitly defined number of parts per maximum of the type.

pub const fn from_percent(x: u32) -> Permill

Converts a percent into Self. Equal to x / 100.

This can be created at compile time.

pub const fn one() -> Permill

See PerThing::one

pub fn is_one(&self) -> bool

See PerThing::is_one.

pub const fn zero() -> Permill

See PerThing::zero.

pub fn is_zero(&self) -> bool

See PerThing::is_zero.

pub const fn deconstruct(self) -> u32

See PerThing::deconstruct.

pub fn square(self) -> Permill

See PerThing::square.

pub fn from_float(x: f64) -> Permill

See PerThing::from_float.

pub fn from_rational_approximation<N>(p: N, q: N) -> Permillwhere N: RationalArg + TryInto<u32> + TryInto<u64>, u32: Into<N>,

๐Deprecated: Use PerThing::from_rational instead

See PerThing::from_rational.

pub fn from_rational<N>(p: N, q: N) -> Permillwhere N: RationalArg + TryInto<u32> + TryInto<u64>, u32: Into<N>,

See PerThing::from_rational.

pub fn int_mul(self, b: u32) -> Permill

Integer multiplication with another value, saturating at 1.

pub fn int_div(self, b: Permill) -> u32

Integer division with another value, rounding down.

pub fn mul_floor<N>(self, b: N) -> Nwhere N: MultiplyArg + UniqueSaturatedInto<u32>, u32: Into<N>,

See PerThing::mul_floor.

pub fn mul_ceil<N>(self, b: N) -> Nwhere N: MultiplyArg + UniqueSaturatedInto<u32>, u32: Into<N>,

See PerThing::mul_ceil.

pub fn saturating_reciprocal_mul<N>(self, b: N) -> Nwhere N: ReciprocalArg + UniqueSaturatedInto<u32>, u32: Into<N>,

See PerThing::saturating_reciprocal_mul.

pub fn saturating_reciprocal_mul_floor<N>(self, b: N) -> Nwhere N: ReciprocalArg + UniqueSaturatedInto<u32>, u32: Into<N>,

See PerThing::saturating_reciprocal_mul_floor.

pub fn saturating_reciprocal_mul_ceil<N>(self, b: N) -> Nwhere N: ReciprocalArg + UniqueSaturatedInto<u32>, u32: Into<N>,

See PerThing::saturating_reciprocal_mul_ceil.

pub fn saturating_div(self, rhs: Permill, r: Rounding) -> Permill

Saturating division. Compute self / rhs, saturating at one if rhs < self.

The rounding method must be specified. e.g.:

let pc = |x| Percent::from_percent(x);
assert_eq!(
	pc(2).saturating_div(pc(3), Down),
	pc(66),
);
assert_eq!(
	pc(1).saturating_div(pc(3), NearestPrefUp),
	pc(33),
);
assert_eq!(
	pc(2).saturating_div(pc(3), NearestPrefDown),
	pc(67),
);
assert_eq!(
	pc(1).saturating_div(pc(3), Up),
	pc(34),
);

impl Permill

pub const fn from_perthousand(x: u32) -> Permill

Converts a percent into Self. Equal to x / 1000.

This can be created at compile time.